requirements.txt
Flask==0.10.1
flask-sqlalchemy
flask-wtf
flask-testing
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
gunicorn==19.3.0
itsdangerous==0.24
six==1.9.0
wsgiref==0.1.2
APScheduler==3.0.0
newrelic
blinker
nose
pylast==1.5.1
spotipy
pyen
fuzzywuzzy